server/requirements.txt
aiofiles
arxiv
python-multipart

# only SQLModel 0.0.14+ supports Pydantic V2
sqlmodel>=0.0.16 
pydantic>=2.0.0
pydantic-settings
psycopg2-binary
psycopg[binary]
pgvector
nest_asyncio

# AI dependencies (LlamaIndex now defaults to V2)
langchain-deepseek
llama-index
llama-index-embeddings-openai
llama-parse
python-dotenv
langchain-core
langchain-openai
langgraph

#Agent Tracing:langGraph Studio
langgraph-cli[inmem] 

#latex
jinja2

#fastapi
fastapi
uvicorn

#testing
httpx

#postgresql async support
asyncpg
langgraph-checkpoint-postgres

#supabase storage
supabase